+++ doubleml/iv_model/__init__.py
@@ -1,0 +1,9 @@
+"""
+The :mod:`doubleml.iv_model` module  implements a variety of instrumental variable models.
+"""
+
+from .pliv import DoubleMLPLIV
+
+__all__ = [
+    "DoubleMLPLIV",
+]
