+++ doubleml/iv_model/iivm.py
@@ -2,12 +2,12 @@ import numpy as np
 from sklearn.utils import check_X_y
 from sklearn.utils.multiclass import type_of_target
 
-from .double_ml import DoubleML
-from .double_ml_data import DoubleMLData
-from .double_ml_score_mixins import LinearScoreMixin
+from ..double_ml import DoubleML
+from ..double_ml_data import DoubleMLData
+from ..double_ml_score_mixins import LinearScoreMixin
 
-from .utils._estimation import _dml_cv_predict, _get_cond_smpls, _dml_tune, _trimm, _normalize_ipw
-from .utils._checks import _check_score, _check_trimming, _check_finite_predictions, _check_is_propensity
+from ..utils._estimation import _dml_cv_predict, _get_cond_smpls, _dml_tune, _trimm, _normalize_ipw
+from ..utils._checks import _check_score, _check_trimming, _check_finite_predictions, _check_is_propensity
 
 
 class DoubleMLIIVM(LinearScoreMixin, DoubleML):
