+++ doubleml/iv_model/tests/_utils_pliv_partial_z_manual.py
@@ -1,7 +1,7 @@
 import numpy as np
 
-from ._utils_boot import boot_manual, draw_weights
-from ._utils import fit_predict, tune_grid_search
+from ...tests._utils_boot import boot_manual, draw_weights
+from ...tests._utils import fit_predict, tune_grid_search
 
 
 def fit_pliv_partial_z(y, x, d, z,
