+++ doubleml/iv_model/tests/test_iivm_no_cross_fit.py
@@ -4,19 +4,16 @@ import math
 
 from sklearn.base import clone
 
-from sklearn.linear_model import LogisticRegression
-from sklearn.ensemble import RandomForestClassifier
+from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor
 
 import doubleml as dml
 
-from ._utils import draw_smpls
+from ...tests._utils import draw_smpls
 from ._utils_iivm_manual import fit_iivm, boot_iivm
 
 
 @pytest.fixture(scope='module',
-                params=[[LogisticRegression(solver='lbfgs', max_iter=250),
-                         LogisticRegression(solver='lbfgs', max_iter=250)],
-                        [RandomForestClassifier(max_depth=2, n_estimators=10),
+                params=[[RandomForestRegressor(max_depth=2, n_estimators=10),
                          RandomForestClassifier(max_depth=2, n_estimators=10)]])
 def learner(request):
     return request.param
@@ -29,31 +26,26 @@ def score(request):
 
 
 @pytest.fixture(scope='module',
-                params=['dml1', 'dml2'])
-def dml_procedure(request):
-    return request.param
-
-
-@pytest.fixture(scope='module',
                 params=[True, False])
 def normalize_ipw(request):
     return request.param
 
 
 @pytest.fixture(scope='module',
-                params=[0.01, 0.05])
-def trimming_threshold(request):
+                params=[1, 2])
+def n_folds(request):
     return request.param
 
 
 @pytest.fixture(scope="module")
-def dml_iivm_classifier_fixture(generate_data_iivm_binary, learner, score, dml_procedure, normalize_ipw, trimming_threshold):
+def dml_iivm_no_cross_fit_fixture(generate_data_iivm, learner, score, normalize_ipw, n_folds):
     boot_methods = ['normal']
-    n_folds = 2
     n_rep_boot = 491
+    dml_procedure = 'dml1'
 
     # collect data
-    (x, y, d, z) = generate_data_iivm_binary
+    data = generate_data_iivm
+    x_cols = data.columns[data.columns.str.startswith('X')].tolist()
 
     # Set machine learning methods for m & g
     ml_g = clone(learner[0])
@@ -61,25 +53,33 @@ def dml_iivm_classifier_fixture(generate_data_iivm_binary, learner, score, dml_p
     ml_r = clone(learner[1])
 
     np.random.seed(3141)
-    obj_dml_data = dml.DoubleMLData.from_arrays(x, y, d, z)
+    obj_dml_data = dml.DoubleMLData(data, 'y', ['d'], x_cols, 'z')
     dml_iivm_obj = dml.DoubleMLIIVM(obj_dml_data,
                                     ml_g, ml_m, ml_r,
                                     n_folds,
                                     dml_procedure=dml_procedure,
                                     normalize_ipw=normalize_ipw,
-                                    trimming_threshold=trimming_threshold)
+                                    apply_cross_fitting=False)
 
     dml_iivm_obj.fit()
 
     np.random.seed(3141)
-    n_obs = len(y)
-    all_smpls = draw_smpls(n_obs, n_folds)
+    y = data['y'].values
+    x = data.loc[:, x_cols].values
+    d = data['d'].values
+    z = data['z'].values
+    if n_folds == 1:
+        smpls = [(np.arange(len(y)), np.arange(len(y)))]
+    else:
+        n_obs = len(y)
+        all_smpls = draw_smpls(n_obs, n_folds)
+        smpls = all_smpls[0]
+        smpls = [smpls[0]]
 
     res_manual = fit_iivm(y, x, d, z,
                           clone(learner[0]), clone(learner[1]), clone(learner[1]),
-                          all_smpls, dml_procedure, score,
-                          normalize_ipw=normalize_ipw,
-                          trimming_threshold=trimming_threshold)
+                          [smpls], dml_procedure, score,
+                          normalize_ipw=normalize_ipw)
 
     res_dict = {'coef': dml_iivm_obj.coef,
                 'coef_manual': res_manual['theta'],
@@ -92,8 +92,9 @@ def dml_iivm_classifier_fixture(generate_data_iivm_binary, learner, score, dml_p
         boot_theta, boot_t_stat = boot_iivm(y, d, z, res_manual['thetas'], res_manual['ses'],
                                             res_manual['all_g_hat0'], res_manual['all_g_hat1'],
                                             res_manual['all_m_hat'], res_manual['all_r_hat0'], res_manual['all_r_hat1'],
-                                            all_smpls, score, bootstrap, n_rep_boot, dml_procedure,
-                                            normalize_ipw=normalize_ipw)
+                                            [smpls], score, bootstrap, n_rep_boot, dml_procedure,
+                                            normalize_ipw=normalize_ipw,
+                                            apply_cross_fitting=False)
 
         np.random.seed(3141)
         dml_iivm_obj.bootstrap(method=bootstrap, n_rep_boot=n_rep_boot)
@@ -106,25 +107,25 @@ def dml_iivm_classifier_fixture(generate_data_iivm_binary, learner, score, dml_p
 
 
 @pytest.mark.ci
-def test_dml_iivm_coef(dml_iivm_classifier_fixture):
-    assert math.isclose(dml_iivm_classifier_fixture['coef'],
-                        dml_iivm_classifier_fixture['coef_manual'],
+def test_dml_iivm_no_cross_fit_coef(dml_iivm_no_cross_fit_fixture):
+    assert math.isclose(dml_iivm_no_cross_fit_fixture['coef'],
+                        dml_iivm_no_cross_fit_fixture['coef_manual'],
                         rel_tol=1e-9, abs_tol=1e-4)
 
 
 @pytest.mark.ci
-def test_dml_iivm_se(dml_iivm_classifier_fixture):
-    assert math.isclose(dml_iivm_classifier_fixture['se'],
-                        dml_iivm_classifier_fixture['se_manual'],
+def test_dml_iivm_no_cross_fit_se(dml_iivm_no_cross_fit_fixture):
+    assert math.isclose(dml_iivm_no_cross_fit_fixture['se'],
+                        dml_iivm_no_cross_fit_fixture['se_manual'],
                         rel_tol=1e-9, abs_tol=1e-4)
 
 
 @pytest.mark.ci
-def test_dml_iivm_boot(dml_iivm_classifier_fixture):
-    for bootstrap in dml_iivm_classifier_fixture['boot_methods']:
-        assert np.allclose(dml_iivm_classifier_fixture['boot_coef' + bootstrap],
-                           dml_iivm_classifier_fixture['boot_coef' + bootstrap + '_manual'],
+def test_dml_iivm_no_cross_fit_boot(dml_iivm_no_cross_fit_fixture):
+    for bootstrap in dml_iivm_no_cross_fit_fixture['boot_methods']:
+        assert np.allclose(dml_iivm_no_cross_fit_fixture['boot_coef' + bootstrap],
+                           dml_iivm_no_cross_fit_fixture['boot_coef' + bootstrap + '_manual'],
                            rtol=1e-9, atol=1e-4)
-        assert np.allclose(dml_iivm_classifier_fixture['boot_t_stat' + bootstrap],
-                           dml_iivm_classifier_fixture['boot_t_stat' + bootstrap + '_manual'],
+        assert np.allclose(dml_iivm_no_cross_fit_fixture['boot_t_stat' + bootstrap],
+                           dml_iivm_no_cross_fit_fixture['boot_t_stat' + bootstrap + '_manual'],
                            rtol=1e-9, atol=1e-4)
