+++ doubleml/tests/test_pliv.py
@@ -10,7 +10,7 @@ from sklearn.ensemble import RandomForestRegressor
 import doubleml as dml
 
 from ._utils import draw_smpls
-from ._utils_pliv_manual import fit_pliv, boot_pliv
+from ..iv_model.tests._utils_pliv_manual import fit_pliv, boot_pliv
 
 
 @pytest.fixture(scope='module',
