+++ doubleml/tests/test_pliv_partial_x_tune.py
@@ -10,7 +10,7 @@ from sklearn.ensemble import RandomForestRegressor
 import doubleml as dml
 
 from ._utils import draw_smpls
-from ._utils_pliv_partial_x_manual import fit_pliv_partial_x, boot_pliv_partial_x, tune_nuisance_pliv_partial_x
+from ..iv_model.tests._utils_pliv_partial_x_manual import fit_pliv_partial_x, boot_pliv_partial_x, tune_nuisance_pliv_partial_x
 
 
 @pytest.fixture(scope='module',
