+++ doubleml/tests/test_pliv_partial_xz_tune.py
@@ -10,7 +10,7 @@ from sklearn.ensemble import RandomForestRegressor
 import doubleml as dml
 
 from ._utils import draw_smpls
-from ._utils_pliv_partial_xz_manual import fit_pliv_partial_xz, boot_pliv_partial_xz, tune_nuisance_pliv_partial_xz
+from ..iv_model.tests._utils_pliv_partial_xz_manual import fit_pliv_partial_xz, boot_pliv_partial_xz, tune_nuisance_pliv_partial_xz
 
 
 @pytest.fixture(scope='module',
