+++ doubleml/tests/test_pliv_partial_z_tune.py
@@ -9,7 +9,7 @@ from sklearn.linear_model import ElasticNet
 import doubleml as dml
 
 from ._utils import draw_smpls
-from ._utils_pliv_partial_z_manual import fit_pliv_partial_z, boot_pliv_partial_z, tune_nuisance_pliv_partial_z
+from ..iv_model.tests._utils_pliv_partial_z_manual import fit_pliv_partial_z, boot_pliv_partial_z, tune_nuisance_pliv_partial_z
 
 
 @pytest.fixture(scope='module',
